requirements.txt
torch==2.2.0
torchvision==0.17.0
numpy==1.24.3
pillow==10.0.0
scikit-learn==1.3.0
matplotlib==3.7.2
pandas==2.0.3
tqdm==4.66.1
opencv-python==4.8.0.76 